added support for bandwidth.
