+++ lib/common/session_model.dart
@@ -117,6 +117,36 @@ class SessionModel extends Model {
     );
   }
 
+  Widget bandwidth(ValueWidgetBuilder<Bandwidth> builder) {
+    if (isMobile()) {
+      return subscribedSingleValueBuilder<Bandwidth>(
+        '/bandwidth',
+        builder: builder,
+        deserialize: (Uint8List serialized) {
+          return Bandwidth.fromBuffer(serialized);
+        },
+      );
+    }
+    final websocket = WebsocketImpl.instance();
+    return ffiValueBuilder<Bandwidth>(
+      'bandwidth',
+      defaultValue: null,
+      onChanges: (setValue) =>
+          sessionModel.listenWebsocket(websocket, "bandwidth", null, (value) {
+            if (value != null) {
+              final Map res = jsonDecode(jsonEncode(value));
+              setValue(Bandwidth.create()
+                ..mergeFromProto3Json({
+                  'allowed': res['mibAllowed'],
+                  'remaining': res['mibUsed'],
+                }));
+            }
+          }),
+      null,
+      builder: builder,
+    );
+  }
+
   Widget developmentMode(ValueWidgetBuilder<bool> builder) {
     if (isMobile()) {
       return subscribedSingleValueBuilder<bool>(
